requirements.txt
Django
psycopg2-binary
Pillow
ipython
python-dotenv
djangorestframework
django-filters
djangorestframework-simplejwt
coverage